Fixed Error with Lab1 Problem4
I apparently thought that A and B would be passed in, but then didn't even give the correct signature for that. This has been fixed to allow an A and B to be specified in the call, if none are provided, it will instead calculate the dot product of the two matrices specified in the problem statement.